Move PositionInitializerFactory
